requirements.txt
Django==1.5
South==1.0
django-auth-ldap==1.2.2
django-bootstrap3==4.11.0
django-crispy-forms==1.4.0
django-json-rpc==0.6.1
django-simple-math-captcha==1.0.5
djangorestframework==2.4.3
django-autocomplete-light==2.0.1
jsonrpc==1.2
python-dateutil==2.2
python-ldap==2.4.17
requests==2.4.3
uuid==1.30
